Change the CARPData data structure.
Instead of storing only unipolar_egm, make CARPData more inline with the Case data structure. It now has a .electric attribute that is the same as that of Case.electric. Currently, only unipolar electrogram files can be added to CARPData, and from this the bipolar electrograms, bipolar voltages, and unipolar voltages are calculated.