Add menubars to each dock widget.
A QMainWindow is set as the widget for each CustomDockWidget, with the BackgroundPlotter being set as the central widget of the QMainWindow. This allows us to add a menubar to each QDockWidget (and later on toolbars too?). The menu will have options for selecting which scalar values to project onto the surface of the mesh.

The menubar of all dock widgets of the active system is coloured blue to signify that this window belongs to the active system. Currently there is no way to set the active system - it is done each time a new system is loaded.

The title of each CustomDockWidget is now bold.

The defalt scalar_bar_args of openep.draw.draw_routines.draw_map have been modified.