Create a table that contains information about each loaded system.
Contians: systme name (can be modified), directory of files for the system, type of system (openCARP or OpenEP), and whether the system is the active system or not (which determine which electrograms will be plotted, and which WOI will be modified).

Can select the active system using radio buttons.

The menubar has been moved from the main GUI QMainWindow to the System DockWidget QMainWindow.