Add functionality to link/unlink views across 3d viewers.
Introduced the concept of a primary viewer. This is simply the first viewer created for a system.

Subsequent viewers are all secondary viewers. Secondary viewers have a button that can be used to link/unlink its view with that of the primary viewer for the system - i.e. when the mesh is rotated in one viewer, it will also rotate in the other.