Interpolate data onto surface when changing the window of interest.
Added a new function openep.case.case_routines.get_sample_indices_within_woi that is used in openep.case.case_routines.calculate_voltage_from_surface.

This correctly accounts for the fact that the window of interest and reference annotaions may be different for different electrograms.

Also updated numpy and scipy versions. Required modifying the matlab reader switch from deprecated scipy functions. Set minimum scipy version to 1.8 to avoid needing to account for the breaking change.

In tests.test_case_routines the bipolar egms must be floats rather than ints, otherwise an error is thrown in get_sample_indices_within_woi when the samples outside the woi are set to np.NaN. They must be set to be NaN to allow for the calculation of the voltage.