Get preference value form the preferences dock rather than the preferences store.
This is because on Windows all values in the store always have type str, whereas on Mac OS and Linux the values take on the correct type if the values in the store are modified. So retrieving the values from the store would need to be handled differently on different os.